Rename test/es6tests.ts to test/ecmascript.ts.
